+++ SandalBotV2/BoardHistory.cpp
@@ -1,6 +1,6 @@
 #include "BoardHistory.h"
 
-void BoardHistory::push(unsigned long long int value) {
+void BoardHistory::push(uint64_t value) {
 	history.push_back(value);
 	mapHistory[value] = history.size() - 1;
 }
@@ -10,11 +10,11 @@ void BoardHistory::pop() {
 	history.pop_back();
 }
 
-unsigned long long int BoardHistory::operator[](const int& index) {
+uint64_t BoardHistory::operator[](const int& index) {
 	return history[index];
 }
 
-bool BoardHistory::operator[](const unsigned long long int& key) {
+bool BoardHistory::operator[](const uint64_t& key) {
 	if (mapHistory.find(key) != mapHistory.end()) {
 		return true;
 	}
